refactor: move api_url to registry module
